MaskRCNN_ROS/include/MaskRCNN/requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow-gpu==1.14.0
keras>=2.0.8
opencv-python
h5py
imageio==2.6.1
imgaug
IPython[all]
pandas
future
